Move GetTopic function out of runc shim
Every shim implementation needs to select a correct publisher topic when posting events, so move it out of Linux-only runc code to the place where other shims can also use it

Otherwise, shims have to copy-paste this code. For example, see runj: https://github.com/samuelkarp/runj/blob/8158e558a361aff1f265a8feed7ad765da5c61b5/containerd/shim.go#L144-L172
